+++ core/7_Content_Classes/Item.lua
@@ -415,7 +415,9 @@ gm.post_script_hook(gm.constants.item_give_internal, function(self, other, resul
     if gm.event_hook_pre_has(args[1].value, gm.constants.ev_destroy, 0, "__actors_holding_item_destroy") then return end
 
     gm.event_hook_pre_add(args[1].value, gm.constants.ev_destroy, 0, "__actors_holding_item_destroy", function(inst)
-        for item_id, _ in pairs(__actors_holding_item[actor_id]) do
+        local t = __actors_holding_item[actor_id]
+        if not t then return end
+        for item_id, _ in pairs(t) do
             __actors_holding_item[item_id][actor_id] = nil
         end
         __actors_holding_item[actor_id] = nil